Move DataSink/SinkResult tests to their correct locations
- TestDataSinkABC → workflow/sinks/tests/test_sinks.py (DataSink lives there)
- TestSinkResult → workflow/schema/sinks/tests/test_sink_configs.py (SinkResult lives there)
- Delete workflow/schema/tests/test_data_sink.py
