+++ python/michelangelo/workflow/schema/sinks/tests/test_sink_configs.py
@@ -79,3 +79,33 @@ class TestInMemorySinkConfig(TestCase):
         """It has a useful repr."""
         cfg = InMemorySinkConfig()
         self.assertIn("InMemorySinkConfig", repr(cfg))
+
+
+class TestSinkResult(TestCase):
+    """Tests for SinkResult frozen dataclass."""
+
+    def test_stores_uri_and_num_records(self):
+        """It stores uri and num_records fields."""
+        from michelangelo.workflow.schema.sinks import SinkResult
+        r = SinkResult(uri="/tmp/data.parquet", num_records=3)
+        self.assertEqual(r.uri, "/tmp/data.parquet")
+        self.assertEqual(r.num_records, 3)
+
+    def test_is_frozen(self):
+        """It raises AttributeError on assignment."""
+        from michelangelo.workflow.schema.sinks import SinkResult
+        r = SinkResult(uri="/tmp/x", num_records=1)
+        with self.assertRaises(AttributeError):
+            r.uri = "/tmp/other"  # type: ignore[misc]
+
+    def test_extra_defaults_to_empty_dict(self):
+        """It defaults extra to an empty dict."""
+        from michelangelo.workflow.schema.sinks import SinkResult
+        r = SinkResult(uri="/tmp/x", num_records=0)
+        self.assertEqual(r.extra, {})
+
+    def test_extra_stores_metadata(self):
+        """It stores arbitrary metadata in extra."""
+        from michelangelo.workflow.schema.sinks import SinkResult
+        r = SinkResult(uri="hive://ml.t", num_records=10, extra={"partitions": 3})
+        self.assertEqual(r.extra["partitions"], 3)
